feat: integrate gossipsub by default
