+++ test/pubsub/floodsub/index.js
@@ -1,0 +1,6 @@
+'use strict'
+
+require('./go2go')
+require('./go2js')
+require('./js2go')
+require('./js2js')
