+++ test/pubsub/index.js
@@ -1,6 +1,4 @@
 'use strict'
 
-require('./go2go')
-require('./go2js')
-require('./js2go')
-require('./js2js')
+require('./gossipsub')
+require('./floodsub')
